Finished the bottom row comparison

+++ twitter.py
@@ -106,7 +106,7 @@ def main():
 
 	# Store the inside value data
 	in_temp = float(sys.argv[1])
-	in_humidity = float(sys.argv[2])
+	in_humidity = int(sys.argv[2])
 	in_heat_index = get_heat_index(in_temp, in_humidity)
 
 	tweet = temp_msg + " in " + org + ", " + city + ", " + region + " " + postal + '\n' +											\
